pyproject.toml
[project]
name = "lol-compare-data-processing"
version = "0.1.0"
description = ""
authors = [
    {name = "Diwas Rai",email = "diwasrai38@yahoo.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas (>=2.3.3,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "isort (>=7.0.0,<8.0.0)",
    "black (>=25.9.0,<26.0.0)",
    "scikit-learn (>=1.7.2,<2.0.0)",
    "joblib (>=1.5.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tools.poetry]
package-mode = false